All todo endpoints jwt protected now

+++ src/todos/todos.service.ts
@@ -14,41 +14,60 @@ export class TodosService {
   }
 
   async findAll(req: any): Promise<Todo[]> {
-    return await this.todoModel
-      .find({ userId: req.user.sub })
-      // .select('-__v -userId');
-      .select('-__v');
+    return await this.todoModel.find({ userId: req.user.sub }).select('-__v');
   }
 
   async findOne(id: string, req: any) {
-    const todo = await this.todoModel.findById(id).select('-__v');
-
-    if (!todo) {
-      throw new NotFoundException();
-    }
-
-    if (todo.userId !== req.user.sub) {
-      throw new NotFoundException('This id does not exist.');
-    }
-    return todo;
-  }
-
-  async update(id: string, updateTodoDto: UpdateTodoDto) {
-    const todo = await this.todoModel
-      .findByIdAndUpdate(id, updateTodoDto, {
-        new: true,
-      })
-      .select('-__v');
-
-    if (!todo) {
-      throw new NotFoundException('This id does not exist.');
-    }
-    return todo;
-  }
-
-  async remove(id: string) {
     try {
+      const todo = await this.todoModel.findById(id).select('-__v');
+
+      if (!todo) {
+        throw new NotFoundException();
+      }
+
+      if (todo.userId !== req.user.sub) {
+        throw new NotFoundException('This id does not exist.');
+      }
+
+      return todo;
+    } catch {
+      throw new NotFoundException('This id does not exist.');
+    }
+  }
+
+  async update(id: string, updateTodoDto: UpdateTodoDto, req: any) {
+    try {
+      const findId = await this.todoModel.findById(id).select('-__v');
+
+      if (findId.userId !== req.user.sub) {
+        throw new NotFoundException('This id does not exist.');
+      }
+
+      const todo = await this.todoModel
+        .findByIdAndUpdate(id, updateTodoDto, {
+          new: true,
+        })
+        .select('-__v');
+
+      if (!todo) {
+        throw new NotFoundException('This id does not exist.');
+      }
+
+      return todo;
+    } catch {
+      throw new NotFoundException('This id does not exist.');
+    }
+  }
+
+  async remove(id: string, req: any) {
+    try {
+      const findId = await this.todoModel.findById(id).select('-__v');
+
+      if (findId.userId !== req.user.sub) {
+        throw new NotFoundException('This id does not exist.');
+      }
       const todo = await this.todoModel.deleteOne({ _id: id });
+
       if (todo.deletedCount === 0) {
         throw new NotFoundException('This id does not exist.');
       }
